File inclusion branch 01 (#7)
First of several Regression branch PR splits. This one includes all the
initial base files for the project, methods, and data. The main methods
files are `pca.py`, `import_data.py`, and `regression.py`.

The plotting base files will be the next commit/branch (not sure how
it's going to work yet) and the pf2 factorization base file will be
included with the other pf2 files.

---------

Co-authored-by: Aaron Meyer <git@asmlab.org>