Inclusion #3: CA data time plots: regression and stripplot of time point distributions (#10)
Includes figure05 and figure06. Linear regression on the CA dataset to
predict time points (if the CA data itself can be used to "predict" the
time it was taken) and a stipplot of the time data distribution to get a
better view of the way to time data is distributed among patients.