Pf2 andrew (#43)
Changes:
  - Upgraded to python>=3.12
- Replaced `parafac2_aoadmm()` with Andrew's `parafac2_nd()` in
`perform_prafac2()`
  - Replaced corresponding test functions in `test_factorization.py`

Bug:
- l1 strength any higher than 1e-5 results in all zero factor matrices
and failed factorization.
- `admm()` in `constrained_parafac()` calls `proximal_operator()` which
performs a `soft_thresholding()` of the C factor matrix using the l1
paramater as the threshold. This results in a matrix of zeros when l1 is
above 1e-5. At the end of iteration 2 of `constrained_parafac()`, `mode
= 2` `constraint_error` is infinity (due to divide by zero) and at the
start of iteration 3, `pseudo_inverse` and `mttkrp` evaluate to 0, which
are then brought into `admm()` as `UtM` and `UtU` as zeros, which
results in factor matrices of `nan` values.